chore: clean up, lint, move write_json to utils, resolve conflicts, update dev deps
