+++ crates/cli/Cargo.toml
@@ -36,4 +36,4 @@ tower-http.workspace = true
 flate2.workspace = true
 
 [dev-dependencies]
-tempfile = "3"
+tempdir.workspace = true
