+++ crates/cli/src/utils.rs
@@ -1,36 +1,13 @@
 use anvil_zksync_config::types::Genesis;
 use anvil_zksync_config::TestNodeConfig;
 use anvil_zksync_core::fork::ForkDetails;
-use anyhow::Context;
-use serde::Serialize;
 use std::fs;
-use std::{
-    fs::File,
-    io::{BufWriter, Write},
-    path::Path,
-};
 
 /// Parses the genesis file from the given path.
 pub fn parse_genesis_file(path: &str) -> Result<Genesis, String> {
     let file_content =
         fs::read_to_string(path).map_err(|err| format!("Failed to read file: {err}"))?;
     serde_json::from_str(&file_content).map_err(|err| format!("Failed to parse JSON: {err}"))
-}
-
-/// Writes the given serializable object as JSON to the specified file path using pretty printing.
-/// Returns an error if the file cannot be created or if serialization/writing fails.
-pub fn write_json_file<T: Serialize>(path: &Path, obj: &T) -> anyhow::Result<()> {
-    let file = File::create(path)
-        .with_context(|| format!("Failed to create file '{}'", path.display()))?;
-    let mut writer = BufWriter::new(file);
-    // Note: intentionally using pretty printing for better readability.
-    serde_json::to_writer_pretty(&mut writer, obj)
-        .with_context(|| format!("Failed to write JSON to '{}'", path.display()))?;
-    writer
-        .flush()
-        .with_context(|| format!("Failed to flush writer for '{}'", path.display()))?;
-
-    Ok(())
 }
 
 /// Updates the configuration from fork details.
