+++ e2e-tests-rust/Cargo.toml
@@ -11,7 +11,7 @@ publish = false
 
 [dependencies]
 alloy-zksync = { git = "https://github.com/popzxc/alloy-zksync.git", rev = "6a2e3cfe41cae9f92dc6672dbd0ce56ccc37e9b3" }
-alloy = { version = "0.6", features = ["full", "rlp", "serde", "sol-types", "getrandom", "provider-anvil-api"] }
+alloy = { version = "0.6", features = ["full", "rlp", "serde", "sol-types", "getrandom", "provider-anvil-api", "json-rpc"] }
 anyhow = "1.0"
 fs2 = "0.4.3"
 tokio = { version = "1", features = ["time", "rt", "process"] }
@@ -23,10 +23,9 @@ reqwest-middleware = { version = "0.4", features = ["json"] }
 serde_json = "1"
 tower = "0.5"
 http = "1.1.0"
+anvil_zksync_core = { path = "../crates/core" }
+tempdir = "0.3.7"
 
 [dev-dependencies]
-serde_json = "1.0"
-anvil_zksync_core = { path = "../crates/core" }
-tempfile = "3"
 
 [workspace] # ignore higher-level workspace
